fix main feed
changed lazy staggered grid on lazy column, use code from photo composables
