+++ myProject/app/src/main/java/com/blblblbl/myapplication/presentation/viewModel/PhotosFragmentViewModel.kt
@@ -1,5 +1,6 @@
 package com.blblblbl.myapplication.presentation.viewModel
 
+import android.util.Log
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import androidx.paging.PagingData
@@ -18,6 +19,7 @@ class PhotosFragmentViewModel @Inject constructor(
 ):ViewModel() {
     lateinit var pagedPhotos: Flow<PagingData<Photo>>
     fun getPhotosFeed(){
+        Log.d("MyLog", "getPhotosFeed()")
         pagedPhotos = getPhotosFeedUseCase.execute(PAGE_SIZE)
     }
     fun changeLike(id: String, bool:Boolean){
