+++ myProject/data/src/main/java/com/blblblbl/myapplication/data/repository/repository_api/utils/RetrofitServices.kt
@@ -48,7 +48,7 @@ class RetrofitServices @Inject constructor(
     )
 
     interface PhotosApi {
-        @Headers("mock:true")
+        //@Headers("mock:true")
         @GET("photos")
         suspend fun getPhotos(
             @Query("page") page: Int,
